mark resolver 分離
